fix: randomization bias depending on first click
The previous code had a subtle bug where there wouldn't be any mines in
the same row or column of the first click. Meanwhile, the intended
feature was to simply prevent the first click to be on a mine.

The condition mineX != buttonX && mineY != buttonY is incorrect.
It should be mineX != buttonX || mineY != buttonY, because if both mineX
and mineY are not equal to buttonX and buttonY respectively, then the mine
will never be planted. This is because the condition requires both mineX
and mineY to be not equal to the corresponding button coordinates, which is
not what is intended. By using the && operator instead of the || operator,
the code is effectively excluding a whole row and column from being
considered for mine placement, and leads to a bias in mine placement.