refactor: simplify button index code
This apparently also fixes a possible bug in non-square mine fields.
This code simplification was pretty obvious but I didn't see it before:
event.GetId() - 10000 IS the button index, we don't have to try and
calculate the X and Y out of it just to calculate the button index back
again.