refactor: use button index trick on right click
The previous code would calculate the button index out of the X and Y
coordinates of the button, however we have received the event! We can
just do:
event.GetId - 10000
And get the button id directly. On a further note, I should really
modify this number to be a variable so I can easily modify it or debug
it in the future.