refactor: simplify sweep code
Apply the same trick again to operate directly on the index instead of
the X and Y coordinates. Apparently, this also re-fixed a bug I fixed a
few commits ago where the buttons clicks would go in incorrect squares
in non-square minefields although I have to test it a bit more to be
sure.