fix: wrong coordinates calculation
Finally managed to squash this bug. OnButtonClicked had a subtle bug in
the calculation of the X and Y coordinates of the button that would only
be apparent when the minefield was non-square. fieldHeight was being
used in a place where fieldWidth should be used instead, however since
in square grids both will be the same the bug won't be manifested.

I thought I fixed this bugs in quite a few previous commits, the reason
for this is I refactored some of the functions to operate on the index
directly instead of bothering with the X and Y coordinates when it's
unnecessary, this meant this bug would not manifest itself in that
specific snippet but didn't actually fix the source of the bug, thus it
kept reappearing randomly in different places.

Ultimately, this is the problem in modelling the minefield as a 1D
array even though it's a 2D field. Theoretically it's possible to
express it in code as bi-dimensional vector and have the compiler
optimize it back to a 1D vector, however that could also complicate
other parts of the code so I'll leave it as it is for now.